chore: refactor: use shared code
